jsonschema

Module output

Implementation of json schema output formats specified in https://json-schema.org/draft/2020-12/json-schema-core.html#rfc.section.12.2

Currently the “flag” and “basic” formats are supported. The “flag” format is idential to the JSONSchema::is_valid method and so is uninteresting. The main contribution of this module is Output::basic. See the documentation of that method for more information.

Structs

Annotations

Annotations associated with an output unit.

ErrorDescription

An error associated with an OutputUnit

Output

The output format resulting from the application of a schema. This can be converted into various representations based on the definitions in https://json-schema.org/draft/2020-12/json-schema-core.html#rfc.section.12.2

OutputUnit

An output unit is a reference to a place in a schema and a place in an instance along with some value associated to that place. For annotations the value will be an Annotations and for errors it will be an ErrorDescription. See the documentation for Output::basic for a detailed example.

Enums

BasicOutput

The “basic” output format. See the documentation for Output::basic for examples of how to use this.