jsonschema::output

Struct OutputUnit

pub struct OutputUnit<T> { /* private fields */ }

An output unit is a reference to a place in a schema and a place in an instance along with some value associated to that place. For annotations the value will be an Annotations and for errors it will be an ErrorDescription. See the documentation for Output::basic for a detailed example.

Implementations

impl<T> OutputUnit<T>

pub const fn keyword_location(&self) -> &JSONPointer

The location in the schema of the keyword

pub const fn absolute_keyword_location(&self) -> &Option<AbsolutePath>

The absolute location in the schema of the keyword. This will be different to keyword_location if the schema is a resolved reference.

pub const fn instance_location(&self) -> &JSONPointer

The location in the instance

impl OutputUnit<Annotations<'_>>

pub fn value(&self) -> Cow<'_, Value>

The annotations found at this output unit

impl OutputUnit<ErrorDescription>

pub const fn error_description(&self) -> &ErrorDescription

The error for this output unit