jsonschema::paths

Struct JSONPointer

pub struct JSONPointer(/* private fields */);

JSON Pointer as a wrapper around individual path components.

Implementations

impl JSONPointer

pub fn into_vec(self) -> Vec<String>

JSON pointer as a vector of strings. Each component is casted to String. Consumes JSONPointer.

pub fn iter(&self) -> Iter<'_, PathChunk>

Return an iterator over the underlying vector of path components.

pub fn last(&self) -> Option<&PathChunk>

Take the last pointer chunk.