Enum jsonschema::error::ValidationErrorKind

pub enum ValidationErrorKind {

    AdditionalItems {
        limit: usize,
    },
    AdditionalProperties {
        unexpected: Vec<String>,
    },
    AnyOf,
    BacktrackLimitExceeded {
        error: Error,
    },
    Constant {
        expected_value: Value,
    },
    Contains,
    ContentEncoding {
        content_encoding: String,
    },
    ContentMediaType {
        content_media_type: String,
    },
    Enum {
        options: Value,
    },
    ExclusiveMaximum {
        limit: Value,
    },
    ExclusiveMinimum {
        limit: Value,
    },
    FalseSchema,
    FileNotFound {
        error: Error,
    },
    Format {
        format: &'static str,
    },
    FromUtf8 {
        error: FromUtf8Error,
    },
    Utf8 {
        error: Utf8Error,
    },
    JSONParse {
        error: Error,
    },
    InvalidReference {
        reference: String,
    },
    InvalidURL {
        error: ParseError,
    },
    MaxItems {
        limit: u64,
    },
    Maximum {
        limit: Value,
    },
    MaxLength {
        limit: u64,
    },
    MaxProperties {
        limit: u64,
    },
    MinItems {
        limit: u64,
    },
    Minimum {
        limit: Value,
    },
    MinLength {
        limit: u64,
    },
    MinProperties {
        limit: u64,
    },
    MultipleOf {
        multiple_of: f64,
    },
    Not {
        schema: Value,
    },
    OneOfMultipleValid,
    OneOfNotValid,
    Pattern {
        pattern: String,
    },
    PropertyNames {
        error: Box<ValidationError<'static>>,
    },
    Required {
        property: Value,
    },
    Schema,
    Type {
        kind: TypeKind,
    },
    UnevaluatedProperties {
        unexpected: Vec<String>,
    },
    UniqueItems,
    UnknownReferenceScheme {
        scheme: String,
    },
    Resolver {
        url: Url,
        error: SchemaResolverError,
    },

}

Kinds of errors that may happen during validation

Variants

AdditionalItems

The input array contain more items than expected.

Fields

limit: usize

AdditionalProperties

Unexpected properties.

Fields

unexpected: Vec<String>

AnyOf

The input value is not valid under any of the schemas listed in the ‘anyOf’ keyword.

BacktrackLimitExceeded

Results from a [fancy_regex::Error::BacktrackLimitExceeded] variant when matching

Fields

error: Error

Constant

The input value doesn’t match expected constant.

Fields

expected_value: Value

Contains

The input array doesn’t contain items conforming to the specified schema.

ContentEncoding

The input value does not respect the defined contentEncoding

Fields

content_encoding: String

ContentMediaType

The input value does not respect the defined contentMediaType

Fields

content_media_type: String

Enum

The input value doesn’t match any of specified options.

Fields

options: Value

ExclusiveMaximum

Value is too large.

Fields

limit: Value

ExclusiveMinimum

Value is too small.

Fields

limit: Value

FalseSchema

Everything is invalid for false schema.

FileNotFound

If the referenced file is not found during ref resolution.

Fields

error: Error

Format

When the input doesn’t match to the specified format.

Fields

format: &'static str

FromUtf8

May happen in contentEncoding validation if base64 encoded data is invalid.

Fields

error: FromUtf8Error

Utf8

Invalid UTF-8 string during percent encoding when resolving happens

Fields

error: Utf8Error

JSONParse

May happen during ref resolution when remote document is not a valid JSON.

Fields

error: Error

InvalidReference

ref value is not valid.

Fields

reference: String

InvalidURL

Invalid URL, e.g. invalid port number or IP address

Fields

error: ParseError

MaxItems

Too many items in an array.

Fields

limit: u64

Maximum

Value is too large.

Fields

limit: Value

MaxLength

String is too long.

Fields

limit: u64

MaxProperties

Too many properties in an object.

Fields

limit: u64

MinItems

Too few items in an array.

Fields

limit: u64

Minimum

Value is too small.

Fields

limit: Value

MinLength

String is too short.

Fields

limit: u64

MinProperties

Not enough properties in an object.

Fields

limit: u64

MultipleOf

When some number is not a multiple of another number.

Fields

multiple_of: f64

Not

Negated schema failed validation.

Fields

schema: Value

OneOfMultipleValid

The given schema is valid under more than one of the schemas listed in the ‘oneOf’ keyword.

OneOfNotValid

The given schema is not valid under any of the schemas listed in the ‘oneOf’ keyword.

Pattern

When the input doesn’t match to a pattern.

Fields

pattern: String

PropertyNames

Object property names are invalid.

Fields

error: Box<ValidationError<'static>>

Required

When a required property is missing.

Fields

property: Value

Schema

Resolved schema failed to compile.

Type

When the input value doesn’t match one or multiple required types.

Fields

kind: TypeKind

UnevaluatedProperties

Unexpected properties.

Fields

unexpected: Vec<String>

UniqueItems

When the input array has non-unique elements.

UnknownReferenceScheme

Reference contains unknown scheme.

Fields

scheme: String

Resolver

Error during schema ref resolution.

Fields

url: Url

The url that was tried to be resolved.

error: SchemaResolverError

The resolution error.